findFollowList(): eliminated duplicate codes, integrated previous 4 function into one
