+++ src/main/java/clone/twitter/repository/FollowMapper.java
@@ -1,7 +1,8 @@
 package clone.twitter.repository;
 
 import clone.twitter.domain.Follow;
-import clone.twitter.domain.User;
+import clone.twitter.repository.dto.UserFollowDto;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
@@ -13,13 +14,7 @@ public interface FollowMapper {
 
     void unfollow(Follow follow);
 
-    List<User> findInitialFollowingList(@Param("followerId") String followerId, @Param("limit") int limit);
-
-    List<User> findNextFollowingList(@Param("followerId") String followerId, @Param("followeeId") String followeeId, @Param("limit") int limit);
-
-    List<User> findInitialFollowedList(@Param("followeeId") String followeeId, @Param("limit") int limit);
-
-    List<User> findNextFollowedList(@Param("followeeId") String followeeId, @Param("followerId") String followerId, @Param("limit") int limit);
+    List<UserFollowDto> findFollowList(@Param("followerId") String followerId, @Param("followeeId") String followeeId, @Param("createdAt") LocalDateTime createdAt, @Param("limit") int limit);
 
     Optional<Follow> findByIds(@Param("followerId") String followerId, @Param("followeeId") String followeeId);
 }
