+++ src/main/java/clone/twitter/repository/FollowRepository.java
@@ -1,7 +1,8 @@
 package clone.twitter.repository;
 
 import clone.twitter.domain.Follow;
-import clone.twitter.domain.User;
+import clone.twitter.repository.dto.UserFollowDto;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,13 +11,7 @@ public interface FollowRepository {
 
     void unfollow(Follow follow);
 
-    List<User> findByFollowerIdOrderByCreatedAtDesc(String followerId);
-
-    List<User> findNextByFollowerIdOrderByCreatedAtDesc(String followerId, String followeeId);
-
-    List<User> findByFolloweeIdOrderByCreatedAtDesc(String followeeId);
-
-    List<User> findNextByFolloweeIdOrderByCreatedAtDesc(String followeeId, String followerId);
+    List<UserFollowDto> findByFollowerIdAndFolloweeIdAndCreatedAtOrderByCreatedAtDesc(String followerId, String followeeId, LocalDateTime createdAt);
 
     Optional<Follow> findByIds(String followerId, String followeeId);
 }
